Examples/ORE-RESTAPI/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
pyyaml>=6.0
rq>=1.16.0
redis>=5.0.0
python-multipart>=0.0.6
aiofiles>=23.0
